feat(AdminDashboard): Jframe
